+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## new
 
+* [ADded] Zeiss `.czi` image format support added
 * [Added] Save training labels image from input annotation
 * [Added] Get all annotation raster with the same color in view
 * [Added] Tiff file format added into reader